rearranged amqp internals
